rename canvasSise to canvas and add bg theme for light and dark
