+++ common/constants/canvas.ts
@@ -1,0 +1,19 @@
+export const CANVAS_SIZE = {
+  width: 4000,
+  height: 2000,
+};
+
+export const CANVAS_BACKGROUND_LIGHT={
+  "bg-1": "#fffff",
+  "bg-2": "#f8f9fa",
+  "bg-3": "#f5faff",
+  "bg-4": "#fffce8",
+  "bg-5": "#fdf8f6",
+}
+export const CANVAS_BACKGROUND_DARK={
+  "bg-1": "#121212",
+  "bg-2": "#161718",
+  "bg-3": "#13171C",
+  "bg-4": "#181605",
+  "bg-5": "#1B1615",
+}
