--- common/constants/canvasSize.ts
@@ -1,4 +1,0 @@
-export const CANVAS_SIZE = {
-  width: 4000,
-  height: 2000,
-};
