+++ modules/room/components/board/Background.tsx
@@ -3,7 +3,7 @@ import { RefObject, useEffect } from "react";
 
 import { motion } from "framer-motion";
 
-import { CANVAS_SIZE } from "@/common/constants/canvasSize";
+import { CANVAS_SIZE } from "@/common/constants/canvas";
 import { useBackground } from "@/common/recoil/background";
 
 import { useBoardPosition } from "../../hooks/useBoardPosition";
